requirements.txt
# python3.10
# tensorflow
# tensorflow-rocm
# pip install --upgrade tensorflow-rocm==2.15.1 -f https://repo.radeon.com/rocm/manylinux/rocm-rel-6.1.3/
# pip install --upgrade tensorflow-rocm==2.16.2 -f https://repo.radeon.com/rocm/manylinux/rocm-rel-6.2.3/
numpy==1.26.4
pillow
